BIMSWARM-Google-Analytics-Dashboard-main/requirements.txt
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.5
geopandas==0.13.2
google-analytics-data==0.17.0
google-api-core==2.11.1
google-auth==2.22.0
googleapis-common-protos==1.59.1
httpcore==0.17.3
httpx==0.24.1
matplotlib==3.7.2
numpy==1.25.1
packaging==23.1
pandas==2.0.3
plotly==5.15.0
dash_mantine_components